Update session & socket management

+++ middleware/RSocket.js
@@ -4,20 +4,24 @@ const { client } = require("../allSocketConnection");
 
 const authSocketMiddleware = async (socket, next) => {
   // since you are sending the token with the query
-  const token = socket.handshake.auth?.token;
-  try {
-    const decoded = jwt.verify(token, config.TOKEN_SECRET);
-    socket.user = decoded;
-  } catch (err) {
-    return next(new Error("NOT AUTHORIZED"));
+  // const token = socket.handshake.auth?.token;
+  // try {
+  //   const decoded = jwt.verify(token, config.TOKEN_SECRET);
+  //   socket.user = decoded;
+  // } catch (err) {
+  //   return next(new Error("NOT AUTHORIZED"));
+  // }
+  const userId = socket.handshake.auth?.userId;
+  
+  if(userId) {
+    const previousSocketId = await client.get(userId.toString());
+    if (previousSocketId) {
+      client.del(userId);
+    }
+    await client.set(userId.toString(), socket.id.toString());
+    console.log(socket.id);
   }
-  const userId = socket.user.id;
-  const previousSocketId = await client.get(userId);
-  if (previousSocketId) {
-    client.del(userId);
-  }
-  await client.set(userId, socket.id);
-
+  
   next();
 };
